+++ .travis.yml
@@ -9,7 +9,7 @@ install:
   - "pip install -r requirements.txt"
   - "python setup.py develop"
 script: 
-  - py.test --pep8 app
-  - py.test -v --cov=app --cov-report term-missing --pep8 test
+  - py.test --pep8 points
+  - py.test -v --cov=points --cov-report term-missing --pep8 test
 after_success:
   coveralls
