+++ README.md
@@ -1,9 +1,9 @@
-Stoic
+Points
 ========
-[![Build Status](https://travis-ci.org/dankolbman/stoic.svg?branch=master)](https://travis-ci.org/dankolbman/stoic)
-[![Coverage Status](https://coveralls.io/repos/github/dankolbman/stoic/badge.svg)](https://coveralls.io/github/dankolbman/stoic)
+[![Build Status](https://travis-ci.org/dankolbman/stoic-points.svg?branch=master)](https://travis-ci.org/dankolbman/stoic-points)
+[![Coverage Status](https://coveralls.io/repos/github/dankolbman/stoic-points/badge.svg)](https://coveralls.io/github/dankolbman/stoic-points)
 
-A Travel Blahg remake
+Points API for Stoic
 
 Installing
 ----------
