+++ requirements.txt
@@ -1,6 +1,5 @@
 click==6.6
 Flask==0.11.1
-Flask-Migrate==2.0.0
 Flask-Script==2.0.5
 Flask-CQLAlchemy==1.2.0
 itsdangerous==0.24
