+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name="stoic",
+    name="points",
     version="0.1.0",
     description="Stoic",
     license="Apache",
